# backend/openapi.yaml
openapi: 3.0.0
info:
  title: 'Swagger Petstore'
  description: 'This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.'
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: 'https://petstore.swagger.io/v3'
    description: 'OpenApi host'
  -
    url: '{schema}://host.dev'
    description: 'OpenApi parameters'
    variables:
      schema:
        enum:
          - https
          - http
        default: https
  -
    url: 'https://virtserver.swaggerhub.com/swagger/Petstore/1.0.0'
    description: 'SwaggerHUB API Mocking'
paths:
...
  /users:
    post:
      summary: 'Adds a new user'
      requestBody:
        content:
          application/json:
            schema:
              properties:
                id:
                  type: string
                name:
                  type: string
              type: object
              example:
                id: 10
                name: 'Jessica Smith'
      responses:
        '200':
          description: OK
...
  '/2.0/repositories/{username}/{slug}':
    get:
      operationId: getRepository
      parameters:
        -
          name: username
          in: path
          required: true
          schema:
            type: string
        -
          name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'The repository'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/repository'
          links:
            repositoryPullRequests:
              $ref: '#/components/links/RepositoryPullRequests'
...
  '/2.0/users/{username}':
    get:
      operationId: getUserByName
      parameters:
        -
          name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'The User'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
          links:
            userRepositories:
              $ref: '#/components/links/UserRepositories'
...
  /user:
    post:
      tags:
        - user
      summary: 'Create user'
      description: 'This can only be done by the logged in user.'
      operationId: createUser
      requestBody:
        description: 'Create user object'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        default:
          description: 'successful operation'
...
  /user/login:
    get:
      tags:
        - user
      summary: 'Logs user into system'
      operationId: loginUser
      parameters:
        -
          name: username
          in: query
          description: 'The user name for login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'successful operation'
          headers:
            X-Rate-Limit:
              description: 'calls per hour allowed by the user'
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: 'date in UTC when token expires'
              schema:
                type: string
                format: datetime
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
        '400':
          description: 'Invalid username/password supplied'
...
  '/user/{username}':
    get:
      summary: 'Get user by user name'
      operationId: getUserByName
      parameters:
        -
          name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'successful operation'
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 'Invalid username supplied'
        '404':
          description: 'User not found'
    put:
      summary: 'Updated user'
      description: 'This can pnly be done by the logged in user.'
      operationId: updateUser
      parameters:
        -
          name: username
          in: path
          description: 'name that to be updated'
          required: true
          schema:
            type: string
      requestBody:
        description: 'Updated user object'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '400':
          description: 'Invalid user supplied'
        '404':
          description: 'User not found'
    delete:
      summary: 'Delete user'
      description: 'This can only be done by the logged in user.'
      operationId: deleteUser
      parameters:
        -
          name: username
          in: path
          description: 'The name that needs to be deleted'
          required: true
          schema:
            type: string
      responses:
        '400':
          description: 'Invalid username supplied'
        '404':
          description: 'User not found'
...
  '/users/{id}':
    put:
      tags:
        - users
      summary: 'Update user entry'
      operationId: updateUser
      parameters:
        -
          name: id
          in: path
          description: UUID
          required: true
        -
          name: user
          in: cookie
          required: true
          schema:
            $ref: '#/components/schemas/User'
      responses:
        '200':
          description: success
        default:
          description: error
    delete:
      tags:
        - users
      summary: 'Remove user entry'
      operationId: deleteUser
      parameters:
        -
          name: id
          in: path
          description: UUID
          required: true
      responses:
        '200':
          description: success
        default:
          description: error
    options:
      responses:
        '200':
          description: 'Some CORS stuff'
        4XX:
          description: 'Some Client Error response'
        5XX:
          description: 'Some Server Error response'
    head:
      responses:
        '200':
          description: 'Only checking if it exists'
components:
  schemas:
...
    repository:
      properties:
        slug:
          type: string
        owner:
          $ref: '#/components/schemas/user'
      type: object
    user:
      properties:
        username:
          type: string
        uuid:
          type: string
      type: object
...
    User:
      title: 'User model'
      description: 'User model'
      properties:
        id:
          title: ID
          description: ID
          type: integer
          format: int64
        username:
          title: Username
          description: Username
          type: string
        firstName:
          title: 'First name'
          description: 'First name'
          type: string
        lastName:
          title: 'Last name'
          description: 'Last name'
          type: string
        email:
          title: Email
          description: Email
          type: string
          format: email
        password:
          title: Password
          description: Password
          type: string
          format: int64
          maximum: 255
        phone:
          title: Phone
          description: Phone
          type: string
          format: msisdn
        userStatus:
          title: 'User status'
          description: 'User status'
          type: integer
          format: int32
      type: object
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: 'http://petstore.swagger.io/oauth/dialog'
          scopes:
            'write:pets': 'modify pets in your account'
            'read:pets': 'read your pets'
    api_key:
      type: apiKey
      name: api_key
      in: header
  links:
    UserRepositories:
      operationId: getRepositoriesByOwner
      parameters:
        username: '$response.body#/username'
...
    RepositoryPullRequests:
      operationId: getPullRequestsByRepository
      parameters:
        username: '$response.body#/owner/username'
        slug: '$response.body#/slug'
...